KvStore: Expose KV Payload for independant use
