+++ py-utils/src/utils/kv_store/kv_store.py
@@ -16,172 +16,10 @@
 # please email opensource@seagate.com or cortx-questions@seagate.com.
 
 import errno
-import re
 import inspect
 from urllib.parse import urlparse
 
-from cortx.utils.kv_store.error import KvStoreError
-
-
-class KvData:
-    """ Base class to represent in memory config data """
-
-    def __init__(self, data, delim='>'):
-        """
-        kvstore will be initialized at the time of load
-        delim is used to split key into hierarchy, e.g. "k1>2" or "k1.k2"
-        """
-        self._data = data
-        if len(delim) > 1:
-            raise KvStoreError(errno.EINVAL, "Invalid delim %s", delim)
-        self._delim = delim
-        self._keys = []
-        self.refresh_keys()
-
-    def get_data(self):
-        return self._data
-
-    def get_keys(self):
-        return self._keys
-
-    def refresh_keys(self):
-        """ Refresh keys. Caller can overrite this """
-        self._refresh_keys(self._data)
-
-    def _refresh_keys(self, data, pkey: str = None):
-        if type(data) == list:
-            for i in range(len(data)):
-                self._keys.append("%s[%d]" % (pkey, i))
-        elif type(data) == dict:
-            for key in data.keys():
-                nkey = key if pkey is None else f"%s%s%s" % (pkey, self._delim,
-                                                             key)
-                if type(data[key]) not in [dict, list]:
-                    self._keys.append(nkey)
-                else:
-                    self._refresh_keys(data[key], nkey)
-        else:
-            raise KvStoreError(errno.ENOSYS, "Cant handle type %s", type(data))
-
-
-class DictKvData(KvData):
-    """ Dict based in memory representation of conf data """
-
-    def __init__(self, data: dict, delim='>'):
-        """
-        Args:
-        data: Dictionary representing keys (hierarchical) and values
-        delim: It is used to split key into hierarchy, e.g. "k1>2" or "k1.k2"
-        """
-        super(DictKvData, self).__init__(data, delim)
-
-    def _set(self, key: str, val: str, data: dict):
-        k = key.split(self._delim, 1)
-
-        # Check if key has index, if so identify index
-        index = None
-        ki = re.split(r'\[([0-9]+)\]', k[0])
-        if len(ki) > 1:
-            if len(ki[0].strip()) == 0:
-                raise KvStoreError(errno.EINVAL, "Invalid key name %s", ki[0])
-            if not ki[1].isnumeric():
-                raise KvStoreError(errno.EINVAL,
-                                   "Invalid key index for the key %s", ki[0])
-            k[0], index = ki[0], int(ki[1])
-
-        if k[0] not in data.keys():
-            data[k[0]] = {}
-        if index is not None:
-            if k[0] not in data.keys() or type(data[k[0]]) != list:
-                data[k[0]] = []
-            # if index is more than list size, extend the list
-            for i in range(len(data[k[0]]), index + 1):
-                data[k[0]].append({})
-            # if this is leaf node of the key
-            if len(k) == 1:
-                data[k[0]][index] = val
-            else:
-                # In case the value is string replace with {}
-                if type(data[k[0]][index]) in [str, list]:
-                    data[k[0]][index] = {}
-                self._set(k[1], val, data[k[0]][index])
-            return
-
-        if len(k) == 1:
-            # if this is leaf node of the key
-            data[k[0]] = val
-        else:
-            # This is not the leaf node of the key, process intermediate node
-            if type(data[k[0]]) in [str, list]:
-                data[k[0]] = {}
-            self._set(k[1], val, data[k[0]])
-
-    def set(self, key: str, val: str):
-        """ Updates the value for the given key in the dictionary """
-        return self._set(key, val, self._data)
-
-    def _get(self, key: str, data: dict) -> str:
-        k = key.split(self._delim, 1)
-        # Check if key has index, if so identify index
-        index = None
-        ki = re.split(r'\[([0-9]+)\]', k[0])
-        if len(ki) > 1:
-            if len(ki[0].strip()) == 0:
-                raise KvStoreError(errno.EINVAL, "Invalid key %s", ki[0])
-
-            if not ki[1].isnumeric():
-                raise KvStoreError(errno.EINVAL,
-                                   "Invalid key index for the key %s", ki[0])
-            k[0], index = ki[0], int(ki[1])
-
-        if k[0] not in data.keys():
-            return None
-        data1 = data[k[0]]
-        if index is not None:
-            if type(data[k[0]]) != list or index >= len(data[k[0]]):
-                return None
-            data1 = data[k[0]][index]
-
-        # if this is leaf node of the key
-        if len(k) == 1: return data1
-
-        # This is not the leaf node of the key, process intermediate node
-        return self._get(k[1], data1)
-
-    def get(self, key: str) -> str:
-        """ Obtain value for the given key """
-        return self._get(key, self._data)
-
-    def _delete(self, key: str, data: dict):
-        k = key.split(self._delim, 1)
-
-        index = None
-        ki = re.split(r'\W+', k[0])
-        if len(ki) > 1:
-            k[0], index = ki[0], int(ki[1])
-
-        if k[0] not in data.keys():
-            return None
-
-        if index is not None:
-            if type(data[k[0]]) != list:
-                raise KvStoreError(errno.EINVAL, "Invalid key %s", key)
-            if index >= len(data[k[0]]):
-                return
-            if len(k) == 1:
-                del data[k[0]][index]
-            else:
-                self._delete(k[1], data[k[0]][index])
-            return
-
-        if len(k) == 1:
-            del data[k[0]]
-        else:
-            self._delete(k[1], data[k[0]])
-
-    def delete(self, key):
-        """ Deletes given set of keys from the dictionary """
-        return self._delete(key, self._data)
+from cortx.utils.kv_store.error import KvError
 
 
 class KvStore:
@@ -212,37 +50,37 @@ class KvStore:
 
     def get(self, keys: list) -> list:
         """ Obtains values of keys. Return list of values. """
-        data = self.load()
+        payload = self.load()
         vals = []
         for key in keys:
-            vals.append(data.get(key))
+            vals.append(payload.get(key))
         return vals
 
     def set(self, keys: list, vals: list):
         """ Updates a given set of keys and values """
         if len(keys) != len(vals):
-            raise KvStoreError(errno.EINVAL, f"Mismatched keys & values %s:%s",\
+            raise KvError(errno.EINVAL, f"Mismatched keys & values %s:%s",\
                 keys, vals)
-        data = self.load()
+        payload = self.load()
         for key, val in zip(keys, vals):
-            data.set(key, val)
-        self.dump(data)
+            payload.set(key, val)
+        self.dump(payload)
 
     def delete(self, keys: list):
         """ Deletes given set of keys from the store """
-        data = self.load()
+        payload = self.load()
         for key in keys:
-            data.delete(key)
-        self.dump(data)
+            payload.delete(key)
+        self.dump(payload)
 
     def load(self, delim='>'):
         """ Loads and returns data from KV storage """
-        raise KvStoreError(errno.ENOSYS, f"%s:load() not implemented", \
+        raise KvError(errno.ENOSYS, f"%s:load() not implemented", \
             type(self).__name__)
 
-    def dump(self, data, delim='>'):
+    def dump(self, payload, delim='>'):
         """ Dumps data onto the KV Storage """
-        raise KvStoreError(errno.ENOSYS, f"%s:dump() not implemented", \
+        raise KvError(errno.ENOSYS, f"%s:dump() not implemented", \
             type(self).__name__)
 
 
@@ -275,4 +113,4 @@ class KvStoreFactory:
                                                         delim)
                 return KvStoreFactory._stores[store_url]
 
-        raise KvStoreError(errno.EINVAL, f"Invalid store type %s", store_type)
+        raise KvError(errno.EINVAL, f"Invalid store type %s", store_type)
